requirements.txt
langchain
faiss-cpu
sentence-transformers
streamlit